+++ clients/src/test/java/org/apache/kafka/common/config/provider/MonitorableConfigProvider.java
@@ -1,0 +1,58 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.common.config.provider;
+
+import org.apache.kafka.common.MetricName;
+import org.apache.kafka.common.config.ConfigData;
+import org.apache.kafka.common.metrics.Measurable;
+import org.apache.kafka.common.metrics.Monitorable;
+import org.apache.kafka.common.metrics.PluginMetrics;
+
+import java.io.IOException;
+import java.util.Map;
+import java.util.Set;
+
+public class MonitorableConfigProvider implements ConfigProvider, Monitorable {
+    public static final String NAME = "name";
+    public static final String DESCRIPTION = "description";
+    protected boolean configured = false;
+
+    @Override
+    public void withPluginMetrics(PluginMetrics metrics) {
+        MetricName metricName = metrics.metricName(NAME, DESCRIPTION, Map.of());
+        metrics.addMetric(metricName, (Measurable) (config, now) -> 123);
+    }
+
+    @Override
+    public ConfigData get(String path) {
+        return null;
+    }
+
+    @Override
+    public ConfigData get(String path, Set<String> keys) {
+        return null;
+    }
+
+    @Override
+    public void close() throws IOException {
+    }
+
+    @Override
+    public void configure(Map<String, ?> configs) {
+        configured = true;
+    }
+}
